+++ src/docstrings.jl
@@ -20,7 +20,9 @@ julia> db = connect(duckdb());
 
 julia> copy_to(db, df, "df_mem");
 
-julia> @chain db_table(db, :df_mem) begin
+julia> df_mem = db_table(db, :df_mem);
+
+julia> @chain t(df_mem) begin
          @select(groups:percent)
          @collect
        end
@@ -39,7 +41,7 @@ julia> @chain db_table(db, :df_mem) begin
    9 │ bb          4      0.9
   10 │ aa          5      1.0
 
-julia> @chain db_table(db, :df_mem) begin
+julia> @chain t(df_mem) begin
          @select(contains("e"))
          @collect
        end
@@ -968,20 +970,30 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
 julia> db = connect(duckdb());
 
 julia> copy_to(db, df, "df_mem");
+
+julia> @chain db_table(db, :df_mem) begin
+        @group_by groups
+        @window_frame(3)
+        @window_order(desc(percent))
+        @mutate(avg = mean(value))
+       #@show_query 
+       end;
 ```
 """
 
 const docstring_window_frame = 
 """
-    @window_frame(sql_query, frame_start::Int, frame_end::Int)
+    @window_frame(sql_query, args...)
 
 Define the window frame for window functions in a SQL query, specifying the range of rows to include in the calculation relative to the current row.
 
 # Arguments
-sql_query: The SQL query to operate on, expected to be an instance of SQLQuery.
-- `frame_start`: The starting point of the window frame. A positive value indicates the start after the current row (FOLLOWING), a negative value indicates before the current row (PRECEDING), and 0 indicates the current row.
-- `frame_end`: The ending point of the window frame. A positive value indicates the end after the current row (FOLLOWING), a negative value indicates before the current row (PRECEDING), and 0 indicates the current row.
-
+- `sqlquery::SQLQuery`: The SQLQuery instance to which the window frame will be applied.
+- `args...`: A variable number of arguments specifying the frame boundaries. These can be:
+    - `from`: The starting point of the frame. Can be a positive or negative integer, 0 or empty. When empty, it will use UNBOUNDED
+    - `to`: The ending point of the frame. Can be a positive or negative integer,  0 or empty. When empty, it will use UNBOUNDED
+    - if only one integer is provided without specifying `to` or `from` it will default to from, and to will be UNBOUNDED.
+    - if no arguments are given, both will be UNBOUNDED
 # Examples 
 ```jldoctest
 julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9], 
@@ -992,6 +1004,36 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
 julia> db = connect(duckdb());
 
 julia> copy_to(db, df, "df_mem");
+
+julia> df_mem = db_table(db, :df_mem);
+
+julia> @chain t(df_mem) begin
+        @group_by groups
+        @window_frame(3)
+        @mutate(avg = mean(percent))
+        #@show_query
+       end;
+
+julia> @chain t(df_mem) begin
+        @group_by groups
+        @window_frame(-3, 3)
+        @mutate(avg = mean(percent))
+        #@show_query
+       end;
+
+julia> @chain t(df_mem) begin
+        @group_by groups
+        @window_frame(to = -3)
+        @mutate(avg = mean(percent))
+        #@show_query
+       end;
+
+julia> @chain t(df_mem) begin
+        @group_by groups
+        @window_frame()
+        @mutate(avg = mean(percent))
+        #@show_query
+       end;
 ```
 """
 
@@ -1037,7 +1079,10 @@ This function establishes a database connection based on the specified backend a
 # Connect to AWS via DuckDB
 # aws_db = connect2(duckdb(), :aws, aws_access_key_id=get(ENV, "AWS_ACCESS_KEY_ID", "access_key"), aws_secret_access_key=get(ENV, "AWS_SECRET_ACCESS_KEY", "secret_access key"), aws_region=get(ENV, "AWS_DEFAULT_REGION", "us-east-1"))
 # Connect to MotherDuck
-# connect(duckdb(), "token") for first connection, vs connect(duckdb(), "md:") for reconnection
+# connect(duckdb(), ""md://..."") for first connection, vs connect(duckdb(), "md:") for reconnection
+# Connect to exisiting database file
+# connect(duckdb(), "path/to/database.duckdb")
+# Open an in-memory database
 julia> db = connect(duckdb())
 DuckDB.Connection(":memory:")
 ```
@@ -1095,7 +1140,7 @@ const docstring_db_table =
     db_table(database, table_name, athena_params, delta = false, iceberg = false)
 
 `db_table` starts the underlying SQL query struct, adding the metadata and table. If paths are passed directly to db_table instead of a 
-name it will not copy it to memory, but rather ready directly from the file.
+name it will not copy it to memory, but rather ready directly from the file. `db_table` only supports direct file paths to a table. It does not support database file paths such as `dbname.duckdb` or `dbname.sqlite`. Such files must be used with `connect first`
 
 # Arguments
 - `database`: The Database or connection object
@@ -1222,7 +1267,7 @@ Shows tables available in database. currently supports DuckDB, databricks, Snowf
 ```jldoctest
 julia> db = connect(duckdb());
 
-julia> show_tables(db); # there are no tables in when first loading so df below is empty.
+julia> show_tables(db);
 ```
 """